feat(web): add spreadsheet-style row/column labels to CSV preview
Render CSV/TSV previews like Excel opens a CSV: a sticky A/B/C column
letter header row and a sticky 1/2/3 row-number gutter, so cell
coordinates like B12 can be located directly.

All rows are now rendered as data rows numbered from 1 instead of
promoting the first CSV row to a styled header, keeping coordinates
consistent with Excel. The grid is sized to the widest row so ragged
CSVs no longer silently drop cells.

Extract colLabel from XlsxPreview into a shared lib/spreadsheet helper.

Closes #17
