+++ web/src/components/workspace/file-preview/CsvPreview.tsx
@@ -1,4 +1,5 @@
 import { ScrollBar } from '@/components/ui/scroll-area'
+import { colLabel } from '@/lib/spreadsheet'
 import * as ScrollAreaPrimitive from '@radix-ui/react-scroll-area'
 import Papa from 'papaparse'
 import { useMemo } from 'react'
@@ -6,18 +7,21 @@ import { useTranslation } from 'react-i18next'
 
 export function CsvPreview({ content, filename }: { content: string; filename: string }) {
   const { t } = useTranslation()
-  const { headers, rows } = useMemo(() => {
+  const { rows, columnCount } = useMemo(() => {
     const ext = filename.split('.').pop()?.toLowerCase()
     const result = Papa.parse<string[]>(content, {
       delimiter: ext === 'tsv' ? '\t' : undefined,
       skipEmptyLines: true,
     })
     const data = result.data
-    if (data.length === 0) return { headers: [], rows: [] }
-    return { headers: data[0], rows: data.slice(1) }
+    // Ragged CSVs are common; size the grid to the widest row so no cell is
+    // silently dropped.
+    let columnCount = 0
+    for (const row of data) if (row.length > columnCount) columnCount = row.length
+    return { rows: data, columnCount }
   }, [content, filename])
 
-  if (headers.length === 0) {
+  if (rows.length === 0 || columnCount === 0) {
     return (
       <div className="flex flex-1 items-center justify-center text-sm text-muted-foreground">
         {t('components.csvPreview.empty')}
@@ -32,12 +36,13 @@ export function CsvPreview({ content, filename }: { content: string; filename: s
           <table className="text-xs border-collapse">
             <thead>
               <tr>
-                {headers.map((h, i) => (
+                <th className="sticky top-0 left-0 z-20 bg-muted border border-border" />
+                {Array.from({ length: columnCount }, (_, ci) => (
                   <th
-                    key={i}
-                    className="sticky top-0 bg-muted px-2 py-1.5 text-left font-medium text-muted-foreground border border-border whitespace-nowrap"
+                    key={colLabel(ci)}
+                    className="sticky top-0 z-10 bg-muted px-2 py-1 text-center font-mono text-[11px] font-medium text-muted-foreground border border-border whitespace-nowrap"
                   >
-                    {h}
+                    {colLabel(ci)}
                   </th>
                 ))}
               </tr>
@@ -45,7 +50,10 @@ export function CsvPreview({ content, filename }: { content: string; filename: s
             <tbody>
               {rows.map((row, ri) => (
                 <tr key={ri} className="hover:bg-muted/50">
-                  {headers.map((_, ci) => (
+                  <td className="sticky left-0 z-10 bg-muted px-2 py-1 text-center font-mono text-[11px] text-muted-foreground/80 border border-border">
+                    {ri + 1}
+                  </td>
+                  {Array.from({ length: columnCount }, (_, ci) => (
                     <td key={ci} className="px-2 py-1 border border-border/50 whitespace-nowrap">
                       {row[ci] ?? ''}
                     </td>
@@ -55,7 +63,7 @@ export function CsvPreview({ content, filename }: { content: string; filename: s
             </tbody>
           </table>
           <div className="mt-2 text-mini text-muted-foreground">
-            {t('components.csvPreview.summary', { rows: rows.length, columns: headers.length })}
+            {t('components.csvPreview.summary', { rows: rows.length, columns: columnCount })}
           </div>
         </div>
       </ScrollAreaPrimitive.Viewport>
